deleting unnecessary files and folders, adding links to navigate between home and project pages
